Reverse ViewPager if we're explicitly in Arabic.
This change is unfortunate, since we literally have to flip the
ViewPager pages (such that the last page is the first page and
so on) and actionbar tabs. There is already a bug filed against
Android (https://code.google.com/p/android/issues/detail?id=56831),
and, when it's fixed, we can clean this up insha'Allah.