Improve Quran page loading.
Introduce a QuranExecutorService, which is a (currently) single
threaded service backed by a priority queue (LIFO). Loading images
now return a future which can be canceled by the fragment when it is
going away.

The main observable result of this change is that if you swipe really
fast (jumping many pages), the newest page now loads much quicker
than it did before.