Update audio manager UI.
Commit cdc9148 was pushed a bit early and showcased an early version of
an audio manager. That commit also included the rxjava and rxandroid
libraries (with which the logic to figure out what files are downloaded
is currently implemented).

The current audio manager does nothing (except for display the number
of complete suras downloaded). Future commits are intended to add to
this functionality insha'Allah.