Stop cutting audiobar in immersive mode. Fix #517.
In immersive mode, we wanted the toolbar and audiobar to both adjust
so as to not be hidden by the navigation buttons, title bar, etc.
Due to the fact that fitSystemWindows stops at the first value set
to true, the toolbar was properly displayed, whereas the audio bar
was not.

This was most notable on tablets, where the navigation buttons would
show on the bottom of the device (thus overlaying the audio bar). As
a bonus, was able to simplify quran page layout.