Handle next and previous properly.
Before, next and previous were flipped for audio (the thinking was that,
due to the text of the Quran being Arabic, that these should flip).
However, recently discovered that the play button (and the previous and
next buttons) are supposed to be global symbols denoting the direction of
the tape. Therefore, for Arabic, the buttons should not be mirrored, nor
should their operations be flipped.

As thus, the next button looks like the play button - it moves the tape
forward. The previous button points in the opposite direction of the
play button and moves backwards (in this case, to the previous verse).