Improve orientation change on tablet. Fixes #518.
The orientation change between a single page and dual pages (portrait
and landscape) on tablet when the "dual pages" option was set was very
jarring. This patch removes the window background in code instead of in
the theme, removing it right away for phone (or when the "dual pages"
option is off on tablet), and delays its removal otherwise.

Having the background while the ViewPager re-initializes and sets its
state makes the transition look a lot less jarring.